### 2024-10-18 #7368 @editorialbot
[PRE REVIEW]: CTSEG: A segment picture quantum impurity solver based on TRIQS
**Submitting author:** @nkavokine (<a href="http://orcid.org/0000-0002-8037-7996">Nikita Kavokine</a>)
**Repository:** https://github.com/TRIQS/ctseg
**Branch with paper.md** (empty if default branch): 3.3.x
**Version:** 3.3.0
