### 2024-10-23 8bfc9a3
correct DOIs
