+++ paper/paper.bib
@@ -81,7 +81,7 @@ volume = {215},
 pages = {128-136},
 year = {2017},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2017.01.003},
+doi = {10.1016/j.cpc.2017.01.003},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465517300036},
 author = {Hiroshi Shinaoka and Emanuel Gull and Philipp Werner},
 keywords = {Quantum impurity problems, Continuous-time impurity solver, Hybridization expansion, Complex hybridization functions, Dynamical mean-field theory},
@@ -98,7 +98,7 @@ number = {4},
 pages = {1280-1286},
 year = {2013},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2012.12.013},
+doi = {10.1016/j.cpc.2012.12.013},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465512004092},
 author = {Hartmut Hafermann and Philipp Werner and Emanuel Gull},
 keywords = {CT-QMC, CT-HYB, DMFT, Dynamical mean-field theory},
@@ -114,7 +114,7 @@ volume = {235},
 pages = {388-399},
 year = {2019},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2018.09.007},
+doi = {10.1016/j.cpc.2018.09.007},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465518303217},
 author = {Markus Wallerberger and Andreas Hausoel and Patrik Gunacker and Alexander Kowalski and Nicolaus Parragh and Florian Goth and Karsten Held and Giorgio Sangiovanni},
 keywords = {(continuous-time) quantum Monte Carlo, Anderson impurity model, Dynamical mean field theory, Green’s functions},
@@ -143,5 +143,8 @@ Program title: w2dynamics Program Files doi: http://dx.doi.org/10.17632/rjs7kmjm
   author = {Kavokine, Nikita and M\"uller, Markus and Geroges, Antoine and Parcollet, Olivier},
   journal = {Phys. Rev. Lett.},
   year = {2024},
-  publisher = {American Physical Society}
+  volume = {133},
+  pages = {016501},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.133.016501}
 }
