### 2024-10-28 #7368 @lucydot
Hi @HugoStrand are you available to review this submission to JOSS? If this is the first time you have reviewed for us you can find more information [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). The review process is quite different from other journals as everything happens here, as a conversation on Github.
