### 2024-11-04 #7425 @editorialbot
[REVIEW]: CTSEG: A segment picture quantum impurity solver based on TRIQS
**Submitting author:** @nkavokine (<a href="http://orcid.org/0000-0002-8037-7996">Nikita Kavokine</a>)
**Repository:** https://github.com/TRIQS/ctseg
**Branch with paper.md** (empty if default branch): 3.3.x
**Version:** v3.3.1
