### 2024-11-22 5d9d008
Fix typo in ctseg.rst
