### 2024-11-27 TRIQS/ctseg_J#14 @HugoStrand
Authors and year deprecated in LICENSE.txt
Please consider updating the author and year in the LICENSE.txt file, see
https://github.com/TRIQS/ctseg/blob/3.3.x/LICENSE.txt#L3

Also this is a requirement for the JOSS review, see openjournals/joss-reviews#7425

See e.g. TPRF for inspiration, https://github.com/TRIQS/tprf/blob/3.3.x/LICENSE.txt#L4

Best, Hugo
