### 2024-11-27 TRIQS/ctseg_J#15 @HugoStrand
Test dimer.py does not test anything
The test `dimer.py` does not check the result with any reference data, see
https://github.com/TRIQS/ctseg/blob/3.3.x/test/python/dimer.py#L113

(Part of openjournals/joss-reviews#7425)
