### 2024-11-27 TRIQS/ctseg_J#16 @HugoStrand
Missing tests for correctness?
Dear TRIQS/ctseg authors,

I have looked over the test suite and among all the tests only `two_orbitals.py` seems to be an actual test for correctness (for the non-interacting case). The other tests seems to be regression tests comparing with previous results obtained from the solver itself. (There is some ctint results used for inputs in `spin_spin.py`).

Would it be possible to test against exact reference data? e.g. using finite ED models that can be reduced to an effective retarded impurity problem?

Best regards,
Hugo

(Part of openjournals/joss-reviews#7425)
