### 2024-11-27 1743b72
fix license files
