+++ LICENSE.txt
@@ -1,6 +1,7 @@
-TRIQS: a Toolbox for Research in Interacting Quantum Systems
+CTSEG: segment picture impurity solver with spin-spin interactions based on TRIQS
 
-Copyright (C) 2014 by P. Seth, I. Krivenko, M. Ferrero, O. Parcollet
+Copyright (C) 2024, N. Kavokine, H. Lu, N. Wentzell, T. Ayral, M. Ferrero, O. Parcollet
+Copyright (C) 2024, The Simons Foundation 
 
 TRIQS is free software: you can redistribute it and/or modify it under the
 terms of the GNU General Public License as published by the Free Software
