### 2024-11-27 TRIQS/ctseg_J#14 @nkavokine
Thank you for pointing this out! Problem fixed now.
