### 2024-11-27 TRIQS/ctseg_J#15 @nkavokine
Dear Hugo, thanks for pointing this out. This is on purpose since for this particular test case, the results of short runs can exhibit large variations across platforms (see comment in the file header). This test is rather meant as a correctness test, since the result can be compared to ED (dimer_pyed.ref.h5). But a meaningful comparison would require a longer run than would be practical for an automatic test, therefore no comparison to a reference is carried out in this test as a default.
