### 2024-11-28 e506d44
[doc] add refs to about
