+++ doc/about.rst
@@ -3,8 +3,49 @@
 About CTSEG
 *************
 
-The CTSEG solver has been written by N. Kavokine and O. Parcollet, with contributions from H. Lu and N. Wentzell. The algorithm has been
-introduced in [J. Otsuki, Phys. Rev. B 87, 125102 (2013)]. 
+The TRIQS/ctseg solver has been written by N. Kavokine and O. Parcollet, with contributions from H. Lu and N. Wentzell. The authors thank T. Ayral and M. Ferrero for sharing their unpublished implementation of CTSEG, which served as an inspiration and a benchmark for the present implementation. 
 
-The authors thank T. Ayral and M. Ferrero for sharing their unpublished implementation of CTSEG, which served as an inspiration 
-and a benchmark for the present implementation. 
+The hybridization expansion algorithm was originally developed in 2006 by Philipp Werner in the segment picture formulation [1] and later generalized to arbitrary local traces [2, 3]. Dynamical interactions was introduced later for density-density [4, 5] and spin-spin interactions [6,7], see Ref. [8] for a review.
+
+References
+************
+
+   | [1] Continuous-Time Solver for Quantum Impurity Models
+   | P. Werner, A. Comanac, L. de' Medici, M. Troyer, A. J. Millis
+   | Phys. Rev. Lett. 97, 076405 (2006)
+   | https://doi.org/10.1103/PhysRevLett.97.076405
+
+   | [2] Hybridization expansion impurity solver: General formulation and application to Kondo lattice and two-orbital models
+   | P. Werner, A. J. Millis
+   | Phys. Rev. B 74, 155107 (2006)
+   | https://doi.org/10.1103/PhysRevB.74.155107
+
+   | [3] Quantum Monte Carlo impurity solver for cluster dynamical mean-field theory and electronic structure calculations with adjustable cluster base
+   | K. Haule
+   | Phys. Rev. B 75, 155113 (2007)
+   | https://doi.org/10.1103/PhysRevB.75.155113
+
+   | [4] Efficient Dynamical Mean Field Simulation of the Holstein-Hubbard Model
+   | P. Werner, A. J. Millis
+   | Phys. Rev. Lett. 99, 146404 (2007)
+   | https://doi.org/10.1103/PhysRevLett.99.146404
+
+   | [5] Dynamical Screening in Correlated Electron Materials
+   | P. Werner, A. J. Millis
+   | Phys. Rev. Lett. 104, 146401 (2010)
+   | https://doi.org/10.1103/PhysRevLett.104.146401
+
+   | [6] Spin-boson coupling in continuous-time quantum Monte Carlo
+   | J. Otsuki
+   | Phys. Rev. B 87, 125102 (2013)
+   | https://doi.org/10.1103/PhysRevB.87.125102
+
+   | [7] Double-expansion impurity solver for multiorbital models with dynamically screened 𝑈 and 𝐽
+   | K. Steiner, Y. Nomura, P. Werner
+   | Phys. Rev. B 92, 115123 (2015)
+   | https://doi.org/10.1103/PhysRevB.92.115123
+
+   | [8] Dynamical screening in correlated electron systems —- from lattice models to realistic materials
+   | P. Werner, M. Casula
+   | J. Phys.: Condens. Matter 28 383001 (2016)
+   | https://iopscience.iop.org/article/10.1088/0953-8984/28/38/383001
