### 2024-11-29 608ad23
[doc] add section on tests
