+++ doc/algorithm_implementation/implementation_notes.rst
@@ -116,3 +116,17 @@ Miscellaneous
   at every move if the code is compiled in debug mode. Checks include the time-ordering of the segments and of the lines and 
   columns of :math:`[\Delta]`, the consistency between the times stored in the configuration and in :math:`[\Delta]`, and the 
   consistency between the :math:`J_{\perp}` lines and the labels stored in the segments. 
+
+Automatic tests
+***************
+
+Several automatic tests are supplied with CTSEG. They run short simulations to assess whether the software functions correctly.
+* ``anderson.py``, ``dynamical_U.py``, ``Jperp.py``, ``spin_spin.py`` and ``multiorb.py`` are non-regression tests, that compare their results to a reference generated by CTSEG itself. 
+  They cover most situations that might be handled by CTSEG: single or multiple orbitals, static, dynamic or spin-spin interactions. 
+  In addition, two tests written directly in C++, ``anderson.cpp`` and ``spin_spin.cpp``, are supplied for developers who may wish to check for non-regression without re-generating the Python wrapper. 
+* ``dimer.py`` may be used to check the correctness of CTSEG on an exactly solvable problem (two impurity sites coupled to two bath sites). The reference data can be found
+  in ``dimer_pyed.ref.h5``. A meaningful comparison to the reference requires a long solver run (for this particular problem, the result can have significant fluctuations between machines for short runs). 
+  Such a long run is not carried out as part of the automatic testing procedure and should be set up by the user if an in-depth assessment of correctness is required. 
+* Additionally, the correctness of CTSEG for problems involving dynamical and spin-spin interaction can be checked by comparing the output of ``spin_spin.py`` to the data in ``ctint.ref.h5``, 
+  generated by a very long CTINT run. CTINT carries out an altogether different expansion to solve the problem: therefore, agreement between CTSEG and CTINT is a good indication for the correctness 
+  of both solvers. 
