### 2024-11-29 TRIQS/ctseg_J#16 @nkavokine
Dear Hugo, this is indeed a pertinent question. While most of the automatic tests are non-regression tests, two of them allow for comparison with (nearly) exact reference data.
- `dimer.py` solves a problem of two impurity sites coupled to two bath sites, for which an ED reference is supplied in `dimer_pyed.ref.h5`.
- `spin_spin.py` solves a single orbital problem with dynamical density-density and spin-spin interactions. The same problem was solved with a very long CTINT run, whose output (found in `ctint.ref.h5`) can be used as a reference for CTSEG. Since the two solvers carry out very different expansions, the agreement of their results is a strong indication for the correctness of both solvers.

In addition, `two_orbitals.py` tests for internal consistency of the solver (the result should not depend on the block structure chosen for the Green's function).

I have now added a section "Automatic tests" to the documentation (https://triqs.github.io/ctseg/latest/algorithm_implementation/implementation_notes.html#automatic-tests), so that the function of the different tests is clear to the user.
