### 2024-12-04 123556b
[doc] add conditions for half-filling #12
