+++ doc/guide/step_by_step.rst
@@ -22,11 +22,9 @@ They can be conveniently supplied as a Python dictionary::
 
 * ``beta`` is the inverse temperature. 
 
-* ``n_tau`` is the number of points of the imaginary time grid on which the input hybridization :math:`\Delta(\tau)` is sampled. 
-    It is also the default number of samples for the measured fermionic two-point functions. 
+* ``n_tau`` is the number of points of the imaginary time grid on which the input hybridization :math:`\Delta(\tau)` is sampled. It is also the default number of samples for the measured fermionic two-point functions. 
 
-* ``n_tau_bosonic`` is the number of points of the imaginary time grid on which the bosonic two-point function inputs (:math:`D_0(\tau)` and :math:`\mathcal{J}_{\perp}(\tau)`) are sampled. 
-    It is also the default number of samples for the measured bosonic two-point functions. 
+* ``n_tau_bosonic`` is the number of points of the imaginary time grid on which the bosonic two-point function inputs (:math:`D_0(\tau)` and :math:`\mathcal{J}_{\perp}(\tau)`) are sampled. It is also the default number of samples for the measured bosonic two-point functions. 
 
 Green's function structure
 --------------------------
@@ -243,6 +241,19 @@ The solver is then accordingly set up as::
 
 The value of ``n_tau_bosonic`` supplied in the ``constr_params`` and the number of points in the :math:`\tau` grids of
 the :math:`D(\tau)` and :math:`J_{\perp}(\tau)` inputs must match. 
+
+Conditions for half-filling
+---------------------------
+
+For a particle-hole symmetric, spin-symmetric single-orbital problem, the following values of the chemical potential correspond to half-filling (assuming that the orbital energy ``eps`` is 0):
+
+* In the absence of dynamical density-density interaction: :math:`\mu = U/2`. 
+
+* In the presence of a dynamical density-density interaction :math:`D_{\sigma \sigma'} (\tau)` and possibly a perpendicular spin-spin interaction :math:`J_{\perp}(\tau)`: 
+  
+.. math::
+    
+    \mu = \frac{U + [D_{\uparrow\uparrow} + D_{\uparrow \downarrow}](i\omega_n = 0)}{2}
 
 Solve parameters
 ----------------
