+++ doc/algorithm_implementation/implementation_notes.rst
@@ -121,6 +121,7 @@ Automatic tests
 ***************
 
 Several automatic tests are supplied with CTSEG. They run short simulations to assess whether the software functions correctly.
+
 * ``anderson.py``, ``dynamical_U.py``, ``Jperp.py``, ``spin_spin.py`` and ``multiorb.py`` are non-regression tests, that compare their results to a reference generated by CTSEG itself. 
   They cover most situations that might be handled by CTSEG: single or multiple orbitals, static, dynamic or spin-spin interactions. 
   In addition, two tests written directly in C++, ``anderson.cpp`` and ``spin_spin.cpp``, are supplied for developers who may wish to check for non-regression without re-generating the Python wrapper. 
