### 2024-12-04 12056e6
[doc] correct typo Sperp_tau
