### 2024-12-19 #7425 @egcpvanloon
In the group, we've been using ctseg the last few months and have tested functionality and performance, mostly in the context of single-orbital impurity models with retarded interactions. Installation and usage are straightforward and the code performs well. I have completed the checklist above (incl. some comments) and support publication in JOSS.
