### 2025-01-02 TRIQS/ctseg_J#19 @HugoStrand
Ergodicity problem c.f. triqs_cthyb (possibly due to lack of segment pair moves)
### Description

When two states that differ by the addition or removal of two electrons are close in energy the hybridization expansion requires double insertion/removals to not run into ergodicity problems, producing incorrect results.

Here is minimal attractive U single band example (in an extreme parmeter regime to exacerbate the problem) where `triqs_ctseg` fails to describe the half filled ground state while `triqs_cthyb` has no problem capturing it. The importance of the double moves in `triqs_cthyb` can be seen comparing the auto correlation time (14 moves with double moves and 450 without). The same issue occurs in multiband models with large Hunds coupling J to Hubbard U ratio, making this an important issue for real materials modelling.

### Steps to Reproduce

```python
from triqs.gf import *
from triqs.operators import n
from triqs_cthyb import Solver as Cthyb
from triqs_ctseg import SolverCore as Ctseg

beta = 10.
U = -40.
mu = U/2

solver_init = dict(gf_struct=[('down', 1), ('up', 1)], beta=beta, n_tau=1024*8)
solver_params = dict(
    h_int=U*n("up", 0)*n("down", 0), h_loc0=-mu * (n("up", 0) + n("down", 0)),
    length_cycle=50, n_warmup_cycles=int(1e4), n_cycles=int(1e5))

iw_mesh = MeshImFreq(beta, 'Fermion', 1024)
delta_iw = GfImFreq(indices=[0], mesh=iw_mesh)
delta_iw << SemiCircular(1.)

ctseg = Ctseg(**solver_init)
cthyb = Cthyb(delta_interface=True, n_iw=1024, **solver_init)
cthyb_no_double = Cthyb(delta_interface=True, n_iw=1024, **solver_init)

for S in [ctseg, cthyb, cthyb_no_double]:
    S.Delta_tau["up"] << Fourier(delta_iw)
    S.Delta_tau["down"] << Fourier(delta_iw)

ctseg.solve(measure_state_hist=True, **solver_params)
cthyb.solve(measure_density_matrix=True, use_norm_as_weight=True, **solver_params)
cthyb_no_double.solve(measure_density_matrix=True, use_norm_as_weight=True,
    move_double=False, # On by default in CTHYB
    **solver_params)

empty = lambda S : S.density_matrix[0][0, 0] if type(S) is Cthyb else S.results.state_hist[0]
full = lambda S : S.density_matrix[1][0, 0] if type(S) is Cthyb else S.results.state_hist[3]

print(f'ctseg : empty {empty(ctseg):2.2E}, full {full(ctseg):2.2E}')
for S in [cthyb, cthyb_no_double]:
    print(f'cthyb : empty {empty(S):2.2E}, full {full(S):2.2E}' + \
          f' auto correlation time {S.auto_corr_time:1.1f} move_double = {S.solve_parameters["move_double"]}')
```

**Expected behavior:** The empty and fully occupied state should have the same probability (close to 0.5)
**Actual behavior:** `triqs_ctseg` gives the empty state near unity probability.

```
ctseg : empty 9.97E-01, full 5.76E-05
cthyb : empty 5.03E-01, full 4.94E-01 auto correlation time 14.2 move_double = True
cthyb : empty 5.11E-01, full 4.84E-01 auto correlation time 453.1 move_double = False
```
