### 2025-01-02 #7425 @HugoStrand
Dear @nkavokine and @lucydot,

I have two open concerns regarding
- the testing and benchmarking for functionality/correctnes, and
- ergodicity problems absent in the `triqs_cthyb` implementation of the CTHYB algorithm.

@nkavokine I understand that it might not be possible to test for correctness in the automated tests. However, as you point out (TRIQS/ctseg#15 and TRIQS/ctseg#16) there are examples in the test suite that could be used to check correctness but only if substantial computational resources were used. In `triqs_cthyb` this has been handled by lifting out these cases into a separate set of manual benchmarks, not part of the automated test suite, that is used to test for correctness in harder and more computationally demanding (corner) cases, see https://github.com/TRIQS/cthyb/tree/3.3.x/benchmark . It would be great if you could do the same in `triqs_ctseg` in particular for the comparison with the ctint code.

Regarding the ergodicity problem TRIQS/ctseg#19, I find it surprising that `triqs_cthyb` is able to do "better" than `triqs_ctseg`. However, I am not certain that the minimal example captures the full complexity of the ergodicity problems in the multiorbital models. (Ferrero and Parcollet know the details.) Perhaps this case is due to the choice of an empty configuration as starting configuration. This particular single band case could probably be solved with a segment inversion move. However, that would not solve the issue in multiband models.

Best regards,
Hugo
