### 2025-01-06 #7425 @lucydot
Thank you @egcpvanloon and @HugoStrand for your work and insights. @nkavokine it is over to you and your team to address the points raised by @HugoStrand. Please let me know if I can clarify anything as editor.
