### 2025-02-12 77ddaed
implement double insert / remove segments
