+++ benchmark/spin_spin/spin_spin.py
@@ -1,5 +1,5 @@
 # Single orbital with dynamical spin-spin interactions. 
-# Can be compared with reference obtained with CTINT (ctint.ref.h5), see benchmark folder
+# Data in spin_spin.ref.h5 is obtained by running this script on 800 cores. 
 from triqs.gf import *
 import triqs.utility.mpi as mpi
 from triqs.gf.descriptors import Function
@@ -54,7 +54,7 @@ solve_params = {
     "h_loc0": -mu * (n("up", 0) + n("down", 0)),
     "length_cycle": 50,
     "n_warmup_cycles": 1000,
-    "n_cycles": 10000,
+    "n_cycles": 1000000,
     "measure_F_tau": True,
     "measure_nn_tau": True,
     "measure_nn_static": True
@@ -63,7 +63,7 @@ solve_params = {
 # Solve
 S.solve(**solve_params)
 
-# Save and compare to reference
+# Save data
 if mpi.is_master_node():
     with h5.HDFArchive("spin_spin.out.h5", 'w') as A:
         A['G_tau'] = S.results.G_tau
@@ -71,5 +71,3 @@ if mpi.is_master_node():
         A['nn_tau'] = S.results.nn_tau
         A['nn'] = S.results.nn_static
         A['densities'] = S.results.densities
-
-    h5diff("spin_spin.out.h5", "spin_spin.ref.h5", precision=1e-9) 
