+++ c++/triqs_ctseg/moves.hpp
@@ -19,6 +19,8 @@
 
 #include "./moves/insert_segment.hpp"
 #include "./moves/remove_segment.hpp"
+#include "./moves/double_insert_segment.hpp"
+#include "./moves/double_remove_segment.hpp"
 #include "./moves/split_segment.hpp"
 #include "./moves/regroup_segment.hpp"
 #include "./moves/move_segment.hpp"
