### 2025-02-12 4a3f6d3
implement double insert / remove segments
