+++ c++/triqs_ctseg/moves/double_insert_segment.cpp
@@ -1,0 +1,216 @@
+// Copyright (c) 2022-2024 Simons Foundation
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You may obtain a copy of the License at
+//     https://www.gnu.org/licenses/gpl-3.0.txt
+//
+// Authors: Nikita Kavokine, Hao Lu, Olivier Parcollet, Nils Wentzell
+
+#include "double_insert_segment.hpp"
+#include "../logs.hpp"
+#include <cmath>
+
+namespace triqs_ctseg::moves {
+
+  double double_insert_segment::attempt() {
+
+    LOG("\n =================== ATTEMPT DOUBLE INSERT ================ \n");
+
+    // ------------ Choice of segment --------------
+    // Select insertion colors
+    int color_0 = rng(config.n_color());
+    int color_1 = rng(config.n_color() - 1);
+    if (color_1 >= color_0) ++color_1; // little trick to select another color
+    colors.assign({color_0, color_1});
+    LOG("Inserting at color ({}, {})", color_0, color_1);
+
+    // For each color, perform a single insert
+    for (auto const &[i, color] : itertools::enumerate(colors)) {
+      auto &sl = config.seglists[color];
+
+      // Select insertion window [tau_left,tau_right]
+      tau_t tau_left = tau_t::beta(), tau_right = tau_t::zero();
+
+      // If sl.empty, no segment, we have the window.
+      // Otherwise, we pick up one segment at random
+      if (not sl.empty()) {
+        if (is_full_line(sl.back())) {
+          LOG("Full line, cannot insert. (color {})", color);
+          return 0;
+        }
+        // Randomly choose one existing segment
+        long seg_idx = rng(sl.size());
+        tau_left     = sl[seg_idx].tau_cdag;                     // tau_left is cdag of this segment
+        tau_right    = sl[modulo(seg_idx + 1, sl.size())].tau_c; // tau_right is c of next segment, possibly cyclic
+      }
+
+      // We now have the insertion window [tau_left,tau_right]
+      // Warning : it can be cyclic !
+      LOG("Insertion window is tau_left = {}, tau_right = {} (color {})", tau_left, tau_right, color);
+      window_length[i] = tau_left - tau_right;
+
+      // Choose two random times in insertion window
+      auto dt1 = tau_t::random(rng, window_length[i]);
+      auto dt2 = tau_t::random(rng, window_length[i]);
+      if (dt1 == dt2) {
+        LOG("Insert_segment: generated equal times. Rejecting (color {})", color);
+        return 0;
+      }
+      // We ensure that dt1 < dt2, unless inserting in empty line because :
+      // 1- in non empty line, we insert a segment in a void so a [c cdag]
+      // 2- in an empty line, we can insert [c cdag] or [cdag c]
+      // BEWARE in the reverse move.
+      if (dt1 > dt2 and not sl.empty()) std::swap(dt1, dt2); // if inserting into an empty line, two ways to insert
+
+      // Here is the segment we propose to insert
[…]
