+++ python/triqs_ctseg/solver_core_desc.py
@@ -190,6 +190,10 @@ c.add_method("""void solve (**solve_params_t)""",
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
 | move_remove_segment           | bool                                 | true                                    | Whether to perform the move remove segment                                                                        |
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
+| move_double_insert_segment    | bool                                 | false                                   | Whether to perform the move double insert segment                                                                 |
++-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
+| move_double_remove_segment    | bool                                 | false                                   | Whether to perform the move double remove segment                                                                 |
++-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
 | move_move_segment             | bool                                 | true                                    | Whether to perform the move move segment                                                                          |
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
 | move_split_segment            | bool                                 | true                                    | Whether to perform the move split segment                                                                         |
@@ -336,6 +340,16 @@ c.add_member(c_name = "move_remove_segment",
              c_type = "bool",
              initializer = """ true """,
              doc = r"""Whether to perform the move remove segment""")
+
+c.add_member(c_name = "move_double_insert_segment",
+             c_type = "bool",
+             initializer = """ false """,
+             doc = r"""Whether to perform the move double insert segment""")
+
+c.add_member(c_name = "move_double_remove_segment",
+             c_type = "bool",
+             initializer = """ false """,
+             doc = r"""Whether to perform the move double remove segment""")
 
 c.add_member(c_name = "move_move_segment",
              c_type = "bool",
