### 2025-02-13 #7425 @HugoStrand
Dear @lucydot,

Thank you for the question. The manuscript makes a strong claim that `ctseg` outperforms `cthyb`, being twice as fast for a single orbital problem, while the `ctseg` solver has the ergodicity problem raised in TRIQS/ctseg#19 that is absent in `cthyb`. So without a fix for this the performance claim only holds in parameter regimes where the ergodicity issue is absent (repulsive interaction). With the additional moves implemented I do believe that all the performance claims in the manuscript will hold for generic models, and the `ctseg` solver will be a robust cornerstone in many future research projects.

My recommendation would be to have patience and let the authors implement the fix, instead of resorting to a half measure. I hope that there is sufficient flexibility in the editorial guidelines to accommodate this delay in the editorial process.

Best regards, Hugo
