### 2025-02-19 #7425 @lucydot
Hi @HugoStrand - thanks for expanding on this. I'm inferring from this that you are happy to wait as a reviewer over a longer timescale than usual - if this is the case, thank you (and please correct me if I am wrong).

An important part of the JOSS process is improving the software through review, so if @nkavokine agrees that the changes are desirable and has time to implement them (?), I am happy with this approach as editor.
