### 2025-02-20 1e75460
add a benchmark folder to show the consistency of move_double
