### 2025-02-26 144e50b
fix the bug when try_insert2 has offdiag block
