### 2025-02-27 TRIQS/ctseg_J#15 @nkavokine
We have now also added this test to a dedicated "benchmark" folder as you suggested.
