### 2025-02-27 TRIQS/ctseg_J#16 @nkavokine
We have now added `dimer.py` and `spin_spin.py` to a dedicated "benchmark" folder as you suggested.
