### 2025-02-28 TRIQS/ctseg_J#19 @nkavokine
We agree that, with the original set of moves proposed by Otsuki, CTSEG runs into ergodicity problems in situations with strong attractive interactions. Double segment insertion/removal moves are indeed a way to restore ergodicity. We have now implemented such moves (see https://triqs.github.io/ctseg/latest/guide/moves.html for details), and we recover half-filling in the above example.

### Input

```
from triqs.gf import *
from triqs.operators import n
#from triqs_cthyb import Solver as Cthyb
from triqs_ctseg import SolverCore as Ctseg

beta = 10.
U = -40.
mu = U/2

solver_init = dict(gf_struct=[('down', 1), ('up', 1)], beta=beta, n_tau=1024*8)
solver_params = dict(
    h_int=U*n("up", 0)*n("down", 0), h_loc0=-mu * (n("up", 0) + n("down", 0)),
    length_cycle=50, n_warmup_cycles=int(1e5), n_cycles=int(1e7), move_double_insert_segment=True, move_double_remove_segment=True)

iw_mesh = MeshImFreq(beta, 'Fermion', 1024)
delta_iw = GfImFreq(indices=[0], mesh=iw_mesh)
delta_iw << SemiCircular(1.)

ctseg = Ctseg(**solver_init)
#cthyb = Cthyb(delta_interface=True, n_iw=1024, **solver_init)
#cthyb_no_double = Cthyb(delta_interface=True, n_iw=1024, **solver_init)

for S in [ctseg]:
    S.Delta_tau["up"] << Fourier(delta_iw)
    S.Delta_tau["down"] << Fourier(delta_iw)

ctseg.solve(measure_state_hist=True, **solver_params)

empty = lambda S : S.results.state_hist[0]
full = lambda S : S.results.state_hist[3]

print(f'ctseg : empty {empty(ctseg):2.2E}, full {full(ctseg):2.2E}')
```
### Output
``` ctseg : empty 4.99E-01, full 4.98E-01```
