### 2025-02-28 #7425 @nkavokine
Dear @HugoStrand ,

Thank you for your extensive review work on our code. I believe that we have now addressed your remaining concerns:

- We have lifted the tests that allow for comparison with reference data to a separate "benchamrk" folder, as you suggested.
- We have implemented double segment insertion/removal moves, which solve [TRIQS/ctseg#19 ](https://github.com/TRIQS/ctseg/issues/19), and should, in general, improve ergodicity for impurities with attractive interactions.

Is there anything else you would like us to address before publication?
