### 2025-03-04 #7425 @HugoStrand
Dear @lucydot and @nkavokine,

Given a new minor release with the modifications from the review process I strongly support publication.

Best, Hugo
