### 2025-03-04 5ed4311
make move_double on by default
