+++ python/triqs_ctseg/parameters_solve_params_t.rst
@@ -31,9 +31,9 @@
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
 | move_remove_segment           | bool                                 | true                                    | Whether to perform the move remove segment                                                                        |
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
-| move_double_insert_segment    | bool                                 | false                                   | Whether to perform the move double insert segment                                                                 |
+| move_double_insert_segment    | bool                                 | true                                    | Whether to perform the move double insert segment                                                                 |
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
-| move_double_remove_segment    | bool                                 | false                                   | Whether to perform the move double remove segment                                                                 |
+| move_double_remove_segment    | bool                                 | true                                    | Whether to perform the move double remove segment                                                                 |
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
 | move_move_segment             | bool                                 | true                                    | Whether to perform the move move segment                                                                          |
 +-------------------------------+--------------------------------------+-----------------------------------------+-------------------------------------------------------------------------------------------------------------------+
