### 2025-03-04 fe0d1a4
make move_double on by default
