### 2025-03-04 3f067d5
correct the doc after switching move_double on
