### 2025-03-04 50ecb64
correct the doc after switching move_double on
