### 2025-03-07 #7425 @nkavokine
Dear @HugoStrand ,

I do not have a strong opinion as to the default behavior of double moves. We have now enabled them by default as per your suggestion.

I am aware of the general TRIQS approach to branch management, but it was not clear how to make it compatible with the JOSS policy that a given branch is under review, and changes in response to reviewer comments should be pushed to that branch. Now the new moves have also been pushed to unstable. The JOSS policy is also that a new minor release is made upon paper acceptance, so we were waiting for that to make a new release with the double moves.

@lucydot , can you please advise on the next steps?
