### 2025-03-10 #7425 @lucydot
Thank you for you time and expertise @HugoStrand, and for your clear responses @nkavokine   - it is wonderful to see software functionality improvements as the result of a JOSS review.

As you state @nkavokine, now that the reviewers and editorial team are happy with the submission, you will now be asked to make a release alongside a couple of other tasks. I will ask editorialbot to summarise these below.
