+++ doc/install.rst
@@ -85,7 +85,7 @@ To use a particular version, go into the directory with the sources, and look at
 
 Checkout the version of the code that you want::
 
-     $ git checkout 3.3.0
+     $ git checkout 3.3.1
 
 and follow steps 2 to 4 above to compile the code.
 
