+++ CMakeLists.txt
@@ -6,7 +6,7 @@ endif()
 
 # ############
 # Define Project
-project(triqs_ctseg VERSION 3.3.0 LANGUAGES CXX)
+project(triqs_ctseg VERSION 3.3.1 LANGUAGES CXX)
 get_directory_property(IS_SUBPROJECT PARENT_DIRECTORY)
 
 # ############
