+++ doc/ChangeLog.md
@@ -2,6 +2,30 @@
 
 # Changelog
 
+## Version 3.3.1
+
+CTSEG version 3.3.1 is the release associated with our JOSS software publication.
+In incorporates various suggestions and improvements from the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/7425).
+
+We thank all contributors: Michel Ferrero, Thomas Hahn, Nikita Kavokine, Hao Lu, Hugo U. R. Strand, Erik van Loon, Nils Wentzell
+
+Find below an itemized list of changes in this release.
+
+### General
+* Implement double insert / remove segments and enable by default
+* Add a benchmark folder to show the consistency of move_double
+* Fix license files
+* Fix issue with assignment to const
+* Add JOSS paper to paper subdirectory
+
+### doc
+* Update benchmarks and add comments
+* Various typo corrections
+* Add references to about page
+* Add a section on tests
+* Add debian package and anaconda install instructions
+
+
 ## Version 3.3.0
 
 This is the initial release of ctseg, a TRIQS based segment picture impurity solver
