### 2025-03-12 4b17298
Add JOSS badge to README.md
