### 2025-03-12 d9f9c2f
Update README.md with JOSS manuscript author list
