+++ ChangeLog.md
@@ -1,7 +1,1 @@
-(changelog)=
-
-# Changelog
-
-## Version 3.2.1
-
-This is the initial release for this project.
+doc/ChangeLog.md
