### 2025-03-24 e432d4b
Correct copyright notices to mention authors and assignees, adjust License.txt accordingly
