### 2025-03-24 13b4b48
Correct copyright notices to mention authors and assignees, adjust License.txt accordingly
