### 2025-03-24 b3375fe
Use minimal GPL copyright header in all source files
