+++ c++/triqs_ctseg/solver_core.cpp
@@ -1,19 +1,7 @@
-// Copyright (c) 2022-2024 Simons Foundation
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You may obtain a copy of the License at
-//     https://www.gnu.org/licenses/gpl-3.0.txt
-//
-// Authors: Nikita Kavokine, Olivier Parcollet, Nils Wentzell
+// Copyright (c) 2022--present, The TRIQS/ctseg Authors and their Assignees
+// This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
+// SPDX-License-Identifier: GPL-3.0-or-later
+// See LICENSE.txt in the root of this distribution for details.
 
 #include <triqs/mc_tools/mc_generic.hpp>
 #include <triqs/utility/callbacks.hpp>
