+++ doc/sphinxext/sphinx_autorun/__init__.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2022--present, The TRIQS/ctseg Authors and their Assignees
+# This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
+# SPDX-License-Identifier: GPL-3.0-or-later
+# See LICENSE.txt in the root of this distribution for details.
+
 # -*- coding: utf-8 -*-
 """
 sphinxcontirb.autorun
