### 2025-03-24 feb92c4
Use minimal GPL copyright header in all source files
