+++ test/python/two_orbitals.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2024--present, The TRIQS/ctseg Authors and their Assignees
+# This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
+# SPDX-License-Identifier: GPL-3.0-or-later
+# See LICENSE.txt in the root of this distribution for details.
+
 # Test for a simple two-orbital system with no interaction. 
 # We consider a hybridization function specified as either 
 # 4 blocks of size 1 or 2 blocks of size 2. Results should be identical (move_move is disabled). 
