+++ test/python/spin_spin.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2022--present, The TRIQS/ctseg Authors and their Assignees
+# This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
+# SPDX-License-Identifier: GPL-3.0-or-later
+# See LICENSE.txt in the root of this distribution for details.
+
 # Single orbital with dynamical spin-spin interactions. 
 # Can be compared with reference obtained with CTINT (ctint.ref.h5), see benchmark folder
 from triqs.gf import *
