### 2025-03-24 5707fa2
Fully specify nda types in four_point.hpp file
