### 2025-03-24 d58ad62
Rename LICENSE.txt to LICENSE and move copyright notice into COPYRIGHT
-Also remove redundant COPYING.txt
