+++ test/python/multiorb.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2024--present, The TRIQS/ctseg Authors and their Assignees
 # This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
 # SPDX-License-Identifier: GPL-3.0-or-later
-# See LICENSE.txt in the root of this distribution for details.
+# See LICENSE in the root of this distribution for details.
 
 # Multi-orbital impurity with diagonal Delta(tau). 
 # Number of orbitals can be changed. 
