### 2025-04-15 4ffc61b
Update copyright information to list institutions explicitly
