+++ test/python/dimer.py
@@ -1,7 +1,8 @@
-# Copyright (c) 2024--present, The TRIQS/ctseg Authors and their Assignees
+# Copyright (c) 2024--present, The Simons Foundation
+# Copyright (c) 2024--present, Max Planck Institute for Polymer Research, Mainz, Germany
 # This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
 # SPDX-License-Identifier: GPL-3.0-or-later
-# See LICENSE.txt in the root of this distribution for details.
+# See LICENSE in the root of this distribution for details.
 
 # Two impurity sites coupled to two bath sites.
 # Delta(tau) has off-diagonal components. Can be compared to ED reference dimer_pyed.ref.h5 (see "benchmark" folder)
