+++ benchmark/dimer/dimer.py
@@ -1,4 +1,5 @@
-# Copyright (c) 2025--present, The TRIQS/ctseg Authors and their Assignees
+# Copyright (c) 2025--present, The Simons Foundation
+# Copyright (c) 2025--present, Max Planck Institute for Polymer Research, Mainz, Germany
 # This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
 # SPDX-License-Identifier: GPL-3.0-or-later
 # See LICENSE in the root of this distribution for details.
