+++ share/replace_and_rename.py
@@ -1,8 +1,3 @@
-# Copyright (c) 2022--present, The TRIQS/ctseg Authors and their Assignees
-# This file is part of TRIQS/ctseg and is licensed under the terms of GPLv3 or later.
-# SPDX-License-Identifier: GPL-3.0-or-later
-# See LICENSE in the root of this distribution for details.
-
 #!/usr/bin/env python3
 
 import sys
