+++ doc/conf.py.in
@@ -53,7 +53,7 @@ rst_epilog = """
 version = version,
 )
 
-copyright = '2022--present The TRIQS/ctseg Authors and their Assignees'
+copyright = '2022--present The Simons Foundation, Max-Planck-Institut for Polymer Research Mainz'
 
 mathjax_path = "https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/MathJax.js?config=default"
 templates_path = ['@CMAKE_CURRENT_SOURCE_DIR@/_templates']
