### 2025-04-24 #7425 @lucydot
@nkavokine  - a couple of suggested changes for the paper:
- fix the citation on line 34
- re-format the citation on line 43 (e.g. `we apply it to...which has previously been studied (Dumitrescu et al, 2021)`).
