### 2025-05-12 #7425 @lucydot
@nkavokine I have just issued a PR with some corrections to the citation formatting (linew 15 and 34).

After merging please check the compiled output with `@editorialbot generate pdf`.
