### 2025-05-20 #7425 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
